+++ filament/src/Renderer.cpp
@@ -292,7 +292,12 @@ void FRenderer::renderJob(ArenaScope& arena, FView& view) {
     if (view.isFrontFaceWindingInverted()) baseRenderFlags |= RenderPass::HAS_INVERSE_FRONT_FACES;
 
     RenderPass::RenderFlags colorRenderFlags = baseRenderFlags;
-    if (view.hasVsm())                     colorRenderFlags |= RenderPass::HAS_VSM;
+    switch (view.getShadowType()) {
+        case ShadowType::PCF:   break;
+        case ShadowType::VSM:   colorRenderFlags |= RenderPass::HAS_VSM;            break;
+        case ShadowType::DPCF:  colorRenderFlags |= RenderPass::HAS_DPCF_OR_PCSS;   break;
+        case ShadowType::PCSS:  colorRenderFlags |= RenderPass::HAS_DPCF_OR_PCSS;   break;
+    }
 
     RenderPass::RenderFlags structureRenderFlags = baseRenderFlags;
     if (view.hasPicking())                 structureRenderFlags |= RenderPass::HAS_PICKING;
@@ -584,7 +589,25 @@ void FRenderer::renderJob(ArenaScope& arena, FView& view) {
 
     // TAA for color pass
     if (taaOptions.enabled) {
-        input = ppm.taa(fg, input, view.getFrameHistory(), taaOptions, colorGradingConfig);
+        input = ppm.taa(fg, input, view.getFrameHistory(),
+                getColorHistory(fg, view.getFrameHistory()), taaOptions, colorGradingConfig);
+    }
+
+    struct ExportColorHistoryData {
+        FrameGraphId<FrameGraphTexture> color;
+    };
+    if (taaOptions.enabled) {
+        fg.addPass<ExportColorHistoryData>("Export color history", [&](FrameGraph::Builder& builder,
+                auto& data) {
+            // We need to use sideEffect here to ensure this pass won't be culled. The "output" of
+            // this pass is going to be used during the next frame as an "import".
+            builder.sideEffect();
+            data.color = builder.sample(input);
+        }, [&view](FrameGraphResources const& resources, auto const& data, backend::DriverApi&) {
+            FrameHistory& frameHistory = view.getFrameHistory();
+            FrameHistoryEntry& current = frameHistory.getCurrent();
+            resources.detach(data.color, &current.color, &current.colorDesc);
+        });
     }
 
     // --------------------------------------------------------------------------------------------
@@ -1241,6 +1264,26 @@ void FRenderer::initializeClearFlags() {
 
     mClearFlags = (mClearOptions.clear ? TargetBufferFlags::COLOR : TargetBufferFlags::NONE)
             | TargetBufferFlags::DEPTH_AND_STENCIL;
+}
+
+FrameGraphId<FrameGraphTexture> FRenderer::getColorHistory(FrameGraph& fg,
+        FrameHistory const& frameHistory) noexcept {
+    // Here we import the previous frame's color output and cache it in the blackboard. This ensures
+    // we only import it once per frame even across multiple calls to getColorHistory.
+    Blackboard& blackboard = fg.getBlackboard();
+
+    if (auto history = blackboard.get<FrameGraphTexture>("colorHistory")) {
+        return history;
+    }
+
+    FrameHistoryEntry const& entry = frameHistory[0];
+    if (UTILS_UNLIKELY(!entry.color.handle)) {
+        return {};
+    }
+    FrameGraphId<FrameGraphTexture> colorHistory = fg.import("Color history", entry.colorDesc,
+            FrameGraphTexture::Usage::SAMPLEABLE, entry.color);
+    blackboard["colorHistory"] = colorHistory;
+    return colorHistory;
 }
 
 // ------------------------------------------------------------------------------------------------