+++ filament/src/details/Renderer.h
@@ -187,6 +187,9 @@ private:
         return clock::now() - getUserEpoch();
     }
 
+    static FrameGraphId<FrameGraphTexture> getColorHistory(FrameGraph& fg,
+            FrameHistory const& frameHistory) noexcept;
+
     // keep a reference to our engine
     FEngine& mEngine;
     FrameSkipper mFrameSkipper;