+++ libs/filamentapp/src/FilamentApp.cpp
@@ -19,7 +19,9 @@
 #include <filamentapp/FilamentApp.h>
 
 #if !defined(WIN32)
-#    include <unistd.h>
+#if defined(FILAMENT_SUPPORTS_WAYLAND)
+#    include <SDL_syswm.h>
+#endif
 #else
 #    include <SDL_syswm.h>
 #    include <utils/unwindows.h>
@@ -532,15 +534,37 @@ FilamentApp::Window::Window(FilamentApp* filamentApp,
         mWidth = w;
         mHeight = h;
     } else {
+
+#if defined(FILAMENT_SUPPORTS_WAYLAND)
+        struct {
+            struct wl_display *display;
+            struct wl_surface *surface;
+        } wayland{};
+
+        SDL_SysWMinfo wmi;
+        SDL_VERSION(&wmi.version);
+        ASSERT_POSTCONDITION(SDL_GetWindowWMInfo(mWindow, &wmi), "SDL version unsupported!");
+        if (wmi.subsystem == SDL_SYSWM_WAYLAND) {
+            wayland.display = wmi.info.wl.display;
+            wayland.surface = wmi.info.wl.surface;
+        }
+        void* nativeWindow = &wayland;
+
+        // Create the Engine after the window in case this happens to be a single-threaded platform.
+        // For single-threaded platforms, we need to ensure that Filament's OpenGL context is
+        // current, rather than the one created by SDL.
+        mFilamentApp->mEngine = Engine::create(config.backend, nullptr, nativeWindow);
+#else
+        void* nativeWindow = ::getNativeWindow(mWindow);
+
         // Create the Engine after the window in case this happens to be a single-threaded platform.
         // For single-threaded platforms, we need to ensure that Filament's OpenGL context is
         // current, rather than the one created by SDL.
         mFilamentApp->mEngine = Engine::create(config.backend);
-
+#endif
         // get the resolved backend
         mBackend = config.backend = mFilamentApp->mEngine->getBackend();
 
-        void* nativeWindow = ::getNativeWindow(mWindow);
         void* nativeSwapChain = nativeWindow;
 
 #if defined(__APPLE__)